refactor(api): extract releases from client.ts
